This is the result after spending Sunday hacking on some prototype code. It has some of the desired game functionality but poorly implemented. It looks like the basic structure of the game may be straightforward but it will take a lot of fine-tuning to make it work well.
The class I made plays around with some other stuff, too. For example,
it goes fullscreen which will be handy to know. Although doing that
introduces some other bugs which will need to be worked out.

There are a ton of bugs in this. The structure is terrible and
ugly. This is just to toy around with some of the basic
mechanics. Let's not actually create an architecture like this!

* Bugs
  
  - On one occasion circle coordinates became NaN and circle disappeared
  - Collision detection is painfully bad
  - Change in momentum calculations are *very* wrong
  - Circles are generated too far from player's circle
  - Many crashes due to concurrent modifications of the list structure