Create the DarkMatter class. This class essentially runs the game. You can currently play a `single player' game but it needs a lot of work. We will have to figure out out to add the networked code. Maybe we can run a server that has a copy of the game and takes the cahnges each are making and sends it out to the players.
Also, it would be worthwhile to subclass Matter with an AIMatter
player so that single player can play against some AI. Programming
that AI will probably not be too hard: something along the lines of
moving towards other player if bigger, away if smaller. Or maybe
something like move towards the biggest blob that is smaller than
itself. In this way it would continually get larger until it won.