Successfully switched data structures. We are now using the HashSet instead of the ArrayList because it will allow easier interfacing with the networking code. Essentially, we will be able to send diffs to and from the network which will make for easier network programming. Slow but easier.
Also changed the way the objects are painted. All the Matter objects
are painted from smallest to largest so that the smaller shapes don't
show up in the middle of a large object. That always looked kind of
weird and now we don't have to worry about it.

It might be worth investigating some mechanism to make the game a bit
faster. It seems slow at times but I'm not sure why. It could be the
game loop. It's pretty simplistic so a more capable one could help.