use the new data structures and the way the objects painted.
change the absorbed area when two matters intersect. The new area is more accuracy.

add the background of centerView file, to check the situation when matter is close to the screen side.

Background.png is temporary picture just for centerView test.

next step, I should to solve the situation when matter is close to screen side, the screen will paint outside of map.